refactor(MB): 重构 ernie4_8k 模型文件
- 将 ernie4_8k.js 重命名为 ERNIE/4.0-8K.js
- 更新文件内部的引用路径
- 修改 main.js 中的模型引用
