+++ src/main.js
@@ -1,3 +1,3 @@
-const Ernie4_8K = require('./MB/ernie4_8k.js');
+const Ernie4_8K = require('./MB/ERNIE/4.0-8K');
 
 module.exports = Ernie4_8K;
